requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
MarkupSafe==2.0.1
google-cloud-pubsub==2.10.0
psycopg2-binary==2.9.1                # For PostgreSQL integration
google-cloud-storage==1.42.3           # For GCS integration
python-dotenv==0.19.2                  # For environment variables (.env)

